+++ ICD.Connect.Conferencing/BYOD/IByodHubDevice.cs
@@ -1,0 +1,14 @@
+﻿using ICD.Common.Properties;
+using ICD.Connect.Devices;
+
+namespace ICD.Connect.Conferencing.BYOD
+{
+	public interface IByodHubDevice
+	{
+		/// <summary>
+		/// Destination device for cameras routed through a BYOD Hub Device.
+		/// </summary>
+		[CanBeNull]
+		IDevice CameraDestinationDevice { get; set; }
+	}
+}